+++ jestSetup.ts
@@ -1,2 +1,5 @@
 import '@testing-library/jest-dom';
+import value from 'node:crypto';
 import 'regenerator-runtime/runtime';
+
+Object.defineProperty(globalThis, 'crypto', { value });