+++ packages/bootstrap/src/BootstrapNotToggle.tsx
@@ -1,5 +1,5 @@
 import type { NotToggleProps } from '@react-querybuilder/ts';
-import { useRef } from 'react';
+import { useId } from 'react';
 
 export const BootstrapNotToggle = ({
   className,
@@ -9,19 +9,19 @@ export const BootstrapNotToggle = ({
   checked,
   disabled,
 }: NotToggleProps) => {
-  const id = useRef(`notToggle-${Math.random()}`);
+  const id = useId();
 
   return (
     <div className={`form-check-inline form-switch ${className}`}>
       <input
-        id={id.current}
+        id={id}
         className="form-check-input"
         type="checkbox"
         onChange={e => handleOnChange(e.target.checked)}
         checked={!!checked}
         disabled={disabled}
       />
-      <label title={title} htmlFor={id.current} className="form-check-label">
+      <label title={title} htmlFor={id} className="form-check-label">
         {label}
       </label>
     </div>