+++ packages/react-querybuilder/src/internal/generateID.ts
@@ -1,1 +1,1 @@
-export const generateID = () => Math.random().toString();
+export const generateID = () => crypto.randomUUID();